Change subpage structure logic
